Small refactor
